feat adjusting enemy connection loosing and ending game
Its adding handling enemy connection loosing, when enemy cant connect to
a game more then for one minute, so we are deciding that enemy escaped
and set appropriate game state, closing connection and so on.
Its also set up final messages button to go to home screen for the first
version of game MVP, next we are planing to add button to go to stats or
to start new game with the same enemy.